parsing: move smtlib parser to its own directory
